requirements.txt
streamlit
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
Pillow
python-dotenv